Add Rakefile
We have decided to run the task from a rakefile to make error handling
easier when the task is run without a date. The rakefile will
automatically run yesterdays date when a date isn't provided, or it can
run with a specific date. We have added a bit of feedback for when no
alerts exist for the date supplied.

co-authored-by: Laura <laura@dxw.com>